Restructured imagehandler, propagate header and WCS.
Split ImageEnsemble into a DataEnsemble and DataHandler objects.
The DataHanlder object has three DataEnsemble objects inside it,
in image_data, shifted_data and stacked_data. This allows the
data to always look the same nomatter which stage you
look at. Each have a .data, .WCS and .header attribute.

The WCS is now carried along and the header gets updated.
The WCS is modified when things are shifted.
The header is updated to contain Shifty-specific keywords, so that
later stages of the code does not need to know what keywords are called,
because they are always called the same thing; only the read-in stage
needs to know what to look for and then puts it in an easy to find place.